Split out chunk-string and add tests
